FINAL UPDATE: added nix flake & oats export target
This will be the final update to the thoughts cli, as it's become far
too inefficient and slow, and has been succeeded by the oats cli
(written in zig).

I've added a nix flake due to it being more convenient.

And more importantly, to aid in the migration to the oats cli instead,
I've added an oats export option to convert all the thoughts stored in
the thoughts-cli database to an oats database instead.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>